v0.2.0
Changed layout when multiple servers are reporting metrics for one
project to a tabbed layout. This allows for easier navigation between
the different servers.

By default, the tabs will rotate every 10 seconds. This can be changed
by setting `?interval=4` in the URL. This will change the interval to 4
seconds between tab changes.